requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pymongo>=4.6.1
python-dotenv>=1.0.0
firebase-admin>=6.4.0
pydantic>=2.6.0
email-validator>=2.1.0
httpx>=0.26.0
python-multipart>=0.0.9
pandas>=2.2.0
openpyxl>=3.1.2
